Travel service

+++ deploy/release.sh
@@ -1,37 +1,32 @@
 #!/bin/bash
 
-# Set the download folder path (modify if needed)
-download_folder="$HOME/Downloads"
-archive_file="dev-admin-e9f11f11d86ec53e.zip"
+# Function to clean code git
+function clean_code() {
+  git reset --hard
+  git clean -df
 
-# Function to extract archive
-function extract_archive() {
-  # Check if archive exists
-  if [ ! -f "$download_folder/$archive_file" ]; then
-    echo "Error: Archive '$archive_file' not found in '$download_folder'"
-    exit 1
-  fi
-
-  # Extract the archive to the root directory
-  unzip -q "$download_folder/$archive_file" -d .
-
-    # (Optional) Remove the archive file after extraction
-    rm -f "$download_folder/$archive_file"
-  echo "Extracted '$archive_file' to root directory."
-}
-
-# Function to clean root directory (excluding .git and deploy)
-function clean_root() {
-  find . ! -path "./.git" ! -path "./.git/*" ! -path "./deploy*" -delete
-  echo "Cleaned root directory (excluding .git and deploy folder)."
+  echo "Cleaned code git."
 }
 
 # Function to develop laravel
 function deployment() {
-  cp deploy/.env .env
-  cp deploy/README.md README.md
-  composer update
-  php artisan migrate:fresh --seed
+  # if .env file dose not exists, copy .env.local to .env
+  # cp deploy/.env.local .env
+  if [ ! -f ".env" ]; then
+    cp deploy/.env.local .env
+  fi
+  # if set argument update -u or --update, update composer
+  # composer update
+  if [ "$1" == "-u" ] || [ "$1" == "--update" ]; then
+    composer update
+  fi
+
+  # if set argument migrate -m or --migrate, migrate database
+  # php artisan migrate:fresh --seed
+  if [ "$1" == "-m" ] || [ "$1" == "--migrate" ]; then
+    php artisan migrate:fresh --seed
+  fi
+
   php artisan key:generate
   php artisan storage:link
 
@@ -39,7 +34,7 @@ function deployment() {
 }
 
 # Main script execution
-clean_root
+clean_code
 extract_archive
 deployment
